requirements.txt
azure-mgmt-resource>=20.0.0
azure-identity>=1.7.1
kubernetes==21.7.0
